Move generateTestDeckCastVoteRecords and getTallyReportResults to @votingworks/utils
These utilities generate CVRs and tally report results from test deck ballot
specs.
