+++ apps/design/backend/src/test_decks.ts
@@ -1,27 +1,13 @@
-import { assert, assertDefined, find, uniqueBy } from '@votingworks/basics';
+import { assertDefined, find } from '@votingworks/basics';
 import { randomUUID as uuid } from 'node:crypto';
 import {
-  Admin,
   BallotStyleId,
   BaseBallotProps,
-  ContestId,
-  Election,
   ElectionDefinition,
-  GridLayout,
-  PrecinctId,
-  Tabulation,
-  getGroupIdFromBallotStyleId,
 } from '@votingworks/types';
 import {
-  combineElectionResults,
-  convertVotesDictToTabulationVotes,
-  filterVotesByContestIds,
-  generateTestDeckBallots,
-  getBallotStyleIdPartyIdLookup,
-  getContestsForPrecinctAndElection,
-  groupMapToGroupList,
-  singlePrecinctSelectionFor,
-  tabulateCastVoteRecords,
+  generateTestDeckCastVoteRecords,
+  getTallyReportResults,
   TestDeckBallot as TestDeckBallotSpec,
 } from '@votingworks/utils';
 import {
@@ -197,196 +183,6 @@ export async function createPrecinctSummaryBallotTestDeck({
   emitProgress?.(ballotSpecs.length);
 
   return await concatenatePdfs(ballotPdfs);
-}
-
-/**
- * In order to generate CVRs per sheet, we want to know how contests are
- * arranged by sheet so we can arrange the votes accordingly.
- */
-interface BallotContestLayout {
-  ballotStyleId: BallotStyleId;
-  contestIdsBySheet: Array<ContestId[]>;
-}
-
-function getBallotContestLayouts(
-  gridLayouts: readonly GridLayout[]
-): BallotContestLayout[] {
-  return gridLayouts.map((gridLayout) => {
-    const { ballotStyleId } = gridLayout;
-    const numSheets = Math.max(
-      ...gridLayout.gridPositions.map((gp) => gp.sheetNumber)
-    );
-    const contestIdsBySheet: BallotContestLayout['contestIdsBySheet'] =
-      Array.from({
-        length: numSheets,
-      }).map(() => []);
-    const oneContestOptionPerContest = uniqueBy(
-      gridLayout.gridPositions,
-      ({ contestId }) => contestId
-    );
-    for (const contestOption of oneContestOptionPerContest) {
-      const { contestId } = contestOption;
-      const { sheetNumber } = contestOption;
-      contestIdsBySheet[sheetNumber - 1].push(contestId);
-    }
-    return {
-      ballotStyleId,
-      contestIdsBySheet,
-    };
-  });
-}
-
-export function generateTestDeckCastVoteRecords(
-  election: Election,
-  options: { includeSummaryBallots: boolean }
-): Tabulation.CastVoteRecord[] {
-  const { includeSummaryBallots = false } = options;
-
-  // Generate HMPB ballot specs
-  const hmpbBallotSpecs: TestDeckBallotSpec[] = generateTestDeckBallots({
-    election,
-    ballotFormat: 'bubble',
-    includeBlankBallots: false,
-    includeOvervotedBallots: false,
-  });
-
-  // Generate summary ballot specs if configured
-  const summaryBallotSpecs: TestDeckBallotSpec[] = includeSummaryBallots
-    ? generateTestDeckBallots({
-        election,
-        ballotFormat: 'summary',
-        includeBlankBallots: false,
-        includeOvervotedBallots: false,
-      })
-    : [];
-
-  const ballotContestLayouts: BallotContestLayout[] = getBallotContestLayouts(
-    assertDefined(election.gridLayouts)
-  );
-
-  const ballotStyleIdPartyIdLookup = getBallotStyleIdPartyIdLookup(election);
-
-  const cvrs: Tabulation.CastVoteRecord[] = [];
-
-  // Process HMPB ballots
-  for (const ballotSpec of hmpbBallotSpecs) {
-    const ballotStyleGroupId = getGroupIdFromBallotStyleId({
-      ballotStyleId: ballotSpec.ballotStyleId,
-      election,
-    });
-    const CVR_ATTRIBUTES = {
-      precinctId: ballotSpec.precinctId,
-      ballotStyleGroupId,
-      partyId: ballotStyleIdPartyIdLookup[ballotStyleGroupId],
-      scannerId: 'test-deck',
-      batchId: 'test-deck',
-      votingMethod: 'precinct',
-    } as const;
-
-    const ballotContestLayout = find(
-      ballotContestLayouts,
-      ({ ballotStyleId }) => ballotStyleId === ballotSpec.ballotStyleId
-    );
-
-    // HMPB ballots may be multiple sheets, so generate a CVR for each sheet
-    for (const [
-      sheetZeroIndex,
-      sheetContestIds,
-    ] of ballotContestLayout.contestIdsBySheet.entries()) {
-      cvrs.push({
-        votes: filterVotesByContestIds({
-          votes: convertVotesDictToTabulationVotes(ballotSpec.votes),
-          contestIds: sheetContestIds,
-        }),
-        card: { type: 'hmpb', sheetNumber: sheetZeroIndex + 1 },
-        ...CVR_ATTRIBUTES,
-      });
-    }
-  }
-
-  // Process summary ballots
-  for (const ballotSpec of summaryBallotSpecs) {
-    const ballotStyleGroupId = getGroupIdFromBallotStyleId({
-      ballotStyleId: ballotSpec.ballotStyleId,
-      election,
-    });
-    const CVR_ATTRIBUTES = {
-      precinctId: ballotSpec.precinctId,
-      ballotStyleGroupId,
-      partyId: ballotStyleIdPartyIdLookup[ballotStyleGroupId],
-      scannerId: 'test-deck',
-      batchId: 'test-deck',
-      votingMethod: 'precinct',
-    } as const;
-
-    // Summary/BMD ballots contain all votes on a single "sheet" (the QR code)
-    cvrs.push({
-      votes: convertVotesDictToTabulationVotes(ballotSpec.votes),
-      card: { type: 'bmd' },
-      ...CVR_ATTRIBUTES,
-    });
-  }
-
-  return cvrs;
-}
-
-/**
- * Builds tally report results from CVRs, optionally filtered to a specific precinct.
- */
-export async function getTallyReportResults(
-  election: Election,
-  cvrs: Tabulation.CastVoteRecord[],
-  precinctId?: PrecinctId
-): Promise<Admin.TallyReportResults> {
-  const contestIds = precinctId
-    ? getContestsForPrecinctAndElection(
-        election,
-        singlePrecinctSelectionFor(precinctId)
-      ).map(({ id }) => id)
-    : election.contests.map(({ id }) => id);
-
-  if (election.type === 'general') {
-    const [electionResults] = groupMapToGroupList(
-      await tabulateCastVoteRecords({
-        election,
-        cvrs,
-      })
-    );
-
-    return {
-      hasPartySplits: false,
-      contestIds,
-      scannedResults: electionResults,
-      cardCounts: electionResults.cardCounts,
-    };
-  }
-
-  // for primaries, we need to get card counts split by party
-  const electionResultsByParty = groupMapToGroupList(
-    await tabulateCastVoteRecords({
-      election,
-      groupBy: { groupByParty: true },
-      cvrs,
-    })
-  );
-
-  const electionResults = combineElectionResults({
-    election,
-    allElectionResults: electionResultsByParty,
-  });
-  const cardCountsByParty: Admin.CardCountsByParty = {};
-  for (const partyElectionResults of electionResultsByParty) {
-    const { partyId } = partyElectionResults;
-    assert(partyId !== undefined && !Tabulation.isNoPartyId(partyId));
-    cardCountsByParty[partyId] = partyElectionResults.cardCounts;
-  }
-
-  return {
-    hasPartySplits: true,
-    cardCountsByParty,
-    scannedResults: electionResults,
-    contestIds,
-  };
 }
 
 export const FULL_TEST_DECK_TALLY_REPORT_FILE_NAME =
